test: Move GivenGrid_And_Margin to runtime tests
